Move defaultclicommands to pkg2
