+++ .gitignore
@@ -1,5 +1,8 @@
+# Binary output file
+bin/
 # Temporary root certificate used during build processes
 asciich_root_ca.crt
 # VSCode related files
 .vscode/
 __debug_*
+asciichgolangpublic
