+++ pkg/defaultclicommands/ansiblecmd/collectioncmd/CreateFileStructureCmd.go
