+++ pkg/defaultclicommands/gitlabcmd/Gitlab.go
@@ -2,8 +2,8 @@ package gitlabcmd
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/asciich/asciichgolangpublic/defaultclicommands/gitlabcmd/gitlabmetricscmd"
-	"github.com/asciich/asciichgolangpublic/defaultclicommands/gitlabcmd/pipelineschedulescmd"
+	"github.com/asciich/asciichgolangpublic/pkg/defaultclicommands/gitlabcmd/gitlabmetricscmd"
+	"github.com/asciich/asciichgolangpublic/pkg/defaultclicommands/gitlabcmd/pipelineschedulescmd"
 	"github.com/asciich/asciichgolangpublic/pkg/tracederrors"
 )
