+++ pkg/defaultclicommands/gitlabcmd/pipelineschedulescmd/PipelineSchedules.go
