+++ pkg/defaultclicommands/loggingcmd/loggingexamplescmd/LogInfo.go
