+++ pkg/defaultclicommands/loggingcmd/loggingexamplescmd/LogInfoMultiline.go
