+++ pkg/defaultclicommands/loggingcmd/loggingexamplescmd/LoggingExamples.go
