+++ pkg/defaultclicommands/monitoringcmd/Monitoring.go
@@ -2,7 +2,7 @@ package monitoringcmd
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/asciich/asciichgolangpublic/defaultclicommands/monitoringcmd/prometheuscmd"
+	"github.com/asciich/asciichgolangpublic/pkg/defaultclicommands/monitoringcmd/prometheuscmd"
 	"github.com/asciich/asciichgolangpublic/pkg/tracederrors"
 )
