+++ pkg/filesutils/README.md
@@ -4,5 +4,5 @@
 
 To run the tests of filesutils use:
 ```bash
-bash -c "cd $(git rev-parse --show-toplevel)/pkg/filesutils && go test -v "
+bash -c "cd $(git rev-parse --show-toplevel) && go test -v ./pkg/filesutils/..."
 ```
